2012/88/EU
Fill text here :
http://eur-lex.europa.eu/legal-content/EN/ALL/?uri=OJ:L:2012:051:TOC

Article 1

1.   The technical specification for interoperability (‘TSI’) relating to the ‘track-side control-command and signalling’ subsystem and ‘on-board control-command and signalling subsystems’ of the trans-European rail system, as set out in Annex III, is adopted.

2.   The TSI set out in Annex III to this Decision shall apply to the track-side control-command and signalling subsystem as described in point 2.3 and to the on-board control-command and signalling subsystem as described in point 2.4 of Annex II to Directive 2008/57/EC.

Article 2

1.   Member States shall ensure that, whenever a national protection control-command system is requested on board rolling stock to run on a given line or part of the trans-European network, this system is included in the list of Class B systems, having equal legal value as annexes to the TSI.

2.   Member States shall ensure that the functionality, performance and interfaces of the Class B systems remain as currently specified, except where modifications are needed to mitigate safety-related flaws in those systems.

Article 3

Each Member State shall notify to the other Member States and to the Commission, with regard to Class B systems and to those issues identified as open points in Appendix G to the TSI adopted by this Decision, within 6 months of the notification of this Decision, the following:

(a)

the list of applicable technical rules;

(b)

the conformity assessment and checking procedures to be used for ensuring that the applicable technical rules are indeed applied;

(c)

the bodies it appoints for carrying out those conformity-assessment and checking procedures.

If these elements have already been notified in the context of Decisions 2006/679/EC and 2006/860/EC this obligation is considered fulfilled.

Article 4

1.   The Commission may grant a derogation to the obligation laid down in Section 7.3.2.4 of Annex III concerning the mandatory equipment of lines with the European Train Control System (ETCS) in the context of EU funded projects (Section 7.3.2.4), when signalling is renewed on short (less than 150 km) and discontinuous sections of a line and provided that ETCS is installed before the earlier of these two dates:

—

5 years after the end of the project,

—

the time by which the section of the line is connected to another ETCS equipped line.

2.   The Member State concerned shall forward a file concerning the project to the Commission. This file shall contain an economic analysis showing that there is a substantial economic and/or technical advantage in putting ERTMS into service at the earlier of the two dates mentioned in paragraph 1 rather than during the course of the EU-funded project.

3.   The Commission shall examine the file sent to it and the measures proposed by the Member State and shall notify the results of its examination to the Committee referred to in Article 29 of Directive 2008/57/EC. If a derogation is granted, the Member State shall ensure that ERTMS is installed before the earlier of the two dates mentioned in paragraph 1.

Article 5

Commission Decision 2011/291/EU of 26 April 2011 concerning a technical specification for interoperability relating to the rolling stock sub-system – ‘Locomotives and Passenger rolling stock’ of the trans-European conventional rail system (5) is amended as follows:

(1)

the second indent under the headline ‘Legislative measures in force’ in Section 1.4 ‘Referenced documents’ of the Annex is replaced by ‘Control-Command and Signalling TSI’;

(2)

Section 4.2.3.3.1 is replaced by Annex I to this Decision;

(3)

Table 10 in Section 4.3.4 is replaced by Annex II to this Decision.

Article 6

The ERA technical document ERTMS/ETCS System Requirement Specification (SRS), document reference ‘subset-026’ in version 3.2.0, dated 22 December 2010, can be used as a basis to tender the equipment of lines with ETCS and to carry out tests, but before the entry into force of the so-called baseline 3, the fitment of trains with baseline 3 cannot be requested.

Article 7

Decisions 2006/679/EC and 2006/860/EC are hereby repealed. Their provisions shall however continue to apply in relation to maintenance of projects authorised in accordance with the TSIs annexed to those Decisions and, unless the applicant requests to apply this Decision, to projects for new, renewed or upgraded subsystems which are at an advanced stage of development or the subject of a contract in course of performance at the date of notification of the present Decision.

Article 8

This Decision shall apply 6 months after its notification to the Member States.

Article 9

This Decision is addressed to the Member States.